release 3.1.147 source code

+++ Services/DataArtsStudio/V1/Model/ShowFactoryFullTextResponse.php
@@ -48,7 +48,7 @@ class ShowFactoryFullTextResponse implements ModelInterface, ArrayAccess
         'limit' => 'int32',
         'offset' => 'int32',
         'searchDetails' => null,
-        'totalHits' => 'int32'
+        'totalHits' => 'int64'
     ];
 
     /**
